backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.15.0
pillow==10.0.1
numpy==1.24.3
streamlit==1.28.0
requests==2.31.0
python-multipart==0.0.6